Use Ajv@7, remove dependency on Ajv
